status=None while delete is in progress
The status of `share` or `share snapshot` became `None` when the
related instance is removed but themselves are already queried.

Update the status to `deleting` if they cannot find the related
instance in DB.

Also, update the status to `deleting` when share delete or
snapshot delete is called in `manager`.

Closes-Bug: #1503390